completion/_make: add completion for shell function generated targets
The gnu-make allows calling shell functions in variable assignment,
which then can be used to generate dynamic target. A simple example
like:
```
TARGETS := $(shell echo 'a b c') $(shell echo 'd')
$(TARGETS):
	@echo "Target $@"
```

Which allows you to use `make a` and so on.

For now, the completion would shown as:
```
> make
$(shell echo 'a b c') $(shell echo 'd')
```

After modification, the completion would be:
```
> make
a b c d
```

Signed-off-by: Ling Wang <lingwang@wcysite.com>